Add FilmCard component
